Small structure changes + other small changes - Rename main.py to os_controller.py - Create main.py as a main file to start the app - Create CallbackManager to store callbacks - Rename some stuff
